Reworked fetching ParkingZone from hasura
