+++ src/main/java/org/opentripplanner/hasura_client/mappers/ParkingZonesMapper.java
@@ -1,11 +1,11 @@
-package org.opentripplanner.updater.vehicle_sharing.parking_zones;
+package org.opentripplanner.hasura_client.mappers;
 
 import com.google.gson.JsonObject;
 import org.geotools.geojson.geom.GeometryJSON;
 import org.locationtech.jts.geom.Geometry;
+import org.opentripplanner.hasura_client.hasura_objects.ParkingZone;
 import org.opentripplanner.routing.core.vehicle_sharing.VehicleType;
-import org.opentripplanner.updater.vehicle_sharing.parking_zones.ParkingZonesApiResponse.Feature;
-import org.opentripplanner.updater.vehicle_sharing.parking_zones.ParkingZonesApiResponse.ParkingZone;
+import org.opentripplanner.updater.vehicle_sharing.parking_zones.GeometryParkingZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,8 +18,7 @@ import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
 
-public class ParkingZonesMapper {
-
+public class ParkingZonesMapper extends HasuraToOTPMapper<ParkingZone, GeometryParkingZone> {
 
     private static final Logger LOG = LoggerFactory.getLogger(ParkingZonesMapper.class);
 
@@ -37,9 +36,9 @@ public class ParkingZonesMapper {
     private List<Geometry> mapToGeometries(List<ParkingZone> parkingZones) {
         return parkingZones.stream()
                 .map(ParkingZone::getArea)
-                .map(ParkingZonesApiResponse.Area::getFeatures)
+                .map(ParkingZone.Area::getFeatures)
                 .flatMap(Collection::stream)
-                .map(Feature::getGeometry)
+                .map(ParkingZone.Feature::getGeometry)
                 .map(this::deserializeGeometry)
                 .filter(Objects::nonNull)
                 .collect(toList());
@@ -73,6 +72,13 @@ public class ParkingZonesMapper {
                 .collect(toList());
     }
 
+    //  TODO Don't know how to implement this method. Need to ask Wiktor
+    @Override
+    protected GeometryParkingZone mapSingleHasuraObject(ParkingZone hasuraObject) {
+        return null;
+    }
+
+    @Override
     public List<GeometryParkingZone> map(List<ParkingZone> parkingZones) {
         return parkingZones.stream()
                 .collect(groupingBy(ParkingZone::getProviderId))
